Extract basic set of annotation functions outside of Scalar package
